# apidocs/openapi/certs.yaml
...
openapi: 3.0.1
info:
  title: SuperMQ Certs service
  description: |
    HTTP API for Certs service
    Some useful links:
    - [The SuperMQ repository](https://github.com/absmach/supermq)
  contact:
    email: info@abstractmachines.fr
  license:
    name: Apache 2.0
    url: https://github.com/absmach/supermq/blob/main/LICENSE
  version: 0.15.1
servers:
  - url: http://localhost:9019
  - url: https://localhost:9019
...
paths:
...
  /{domainID}/certs/{clientID}:
    delete:
      operationId: revokeCert
      summary: Revokes a certificate
      description: |
        Revokes a certificate for a given client ID.
      tags:
        - certs
      parameters:
        - $ref: "auth.yaml#/components/parameters/DomainID"
        - $ref: "#/components/parameters/ClientID"
      responses:
        "200":
          $ref: "#/components/responses/RevokeRes"
        "400":
          description: Failed due to malformed query parameters.
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Failed to perform authorization over the entity.
        "404":
          description: |
            Failed to revoke corresponding certificate.
        "422":
          description: Database can't process request.
        "500":
          $ref: "#/components/responses/ServiceError"
...
components:
  parameters:
    ClientID:
      name: clientID
      description: Client ID
      in: path
      schema:
        type: string
        format: uuid
      required: true
...
  schemas:
...
    Revoke:
      type: object
      properties:
        revocation_time:
          type: string
          description: Certificate revocation time
...
  responses:
    ServiceError:
      description: Unexpected server-side error occurred.
...
    RevokeRes:
      description: Certificate revoked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Revoke"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        * Users access: "Authorization: Bearer <user_token>"
security:
  - bearerAuth: []